controllers: quadpid: use double instead of uint16_t for tresholds
This allows fine grain tuning of tresholds

+++ controllers/quadpid/include/quadpid.hpp
@@ -64,15 +64,15 @@ typedef struct {
     PID_t linear_pose_pid;                      /**< Linear pose Kp, Ki, Kd */
     PID_t angular_pose_pid;                     /**< Angular pose Kp, Ki, Kd */
 
-    uint16_t min_distance_for_angular_switch;   /**< Distance approximation to
+    double min_distance_for_angular_switch;   /**< Distance approximation to
                                                      switch to angular
                                                      correction */
 
-    uint16_t min_angle_for_pose_reached;        /**< Angle approximation to
+    double min_angle_for_pose_reached;        /**< Angle approximation to
                                                      switch to position reached
                                                      state */
 
-    uint16_t min_angle_for_target_orientation;  /**< Angle approximation to
+    double min_angle_for_target_orientation;  /**< Angle approximation to
                                                      switch to linear
                                                      correction */
